Create README.md
Call  me coKe (it's how spanyards pronouns Jorge when they are children) as a coke can, not the bad coke. I am glad to attend you anytime if you have simpathy for the devil (I listen to loads of music) Almost never been abroad and speak mixed american/british english because of the hiphop.